CA-404370: enable NBD client only after completing handshake
Previously the client was being enabled as part of
`tapdisk_nbdserver_newclient_fd_new_fixed` after calling
`tapdisk_nbdserver_new_protocol_handshake`. This erroneously assumed
that protocol handshake was complete by the time that function
returned, which is not the case as it simply adds the fd to an event
for the scheduler. Protocol handshake is only complete when we reach
the end of the `tapdisk_nbdserver_handshake_cb` callback and enabling
the client before this means we have two events registered for the
same fd.

This was causing quicktest to fail after blktap2 was removed and NBD
is used for all control domain access to the I/O datapath.

Signed-off-by: Mark Syms <mark.syms@cloud.com>