tapdisk: replace signals handling by signalfd
With libqcow2, each of its threads create an signalfd to handle signals
at the thread level. Implement the same signalfd handler for tapdisk
thread to direct signals.

No functional change are expected, but to achieve this, we would have
had to remove SIGINT from the managed signals in order for GDB to work.

Signed-off-by: Anthoine Bourgeois <anthoine.bourgeois@vates.tech>